# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "o365Replicator",
        "description": "A mock Microsoft 365 email provisioning service for POC/demo use. Saviynt fires a new-hire webhook at `/api/v1/provision`; the service generates an email address, stores a mailbox record, returns it synchronously, and (optionally) writes it back to Saviynt.",
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/v1/config": {
            "get": {
                "tags": [
                    "config"
                ],
                "summary": "Read Config",
                "operationId": "read_config_api_v1_config_get",
                "parameters": [
                    {
                        "name": "x-api-key",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-Api-Key"
                        }
                    },
                    {
                        "name": "authorization",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Authorization"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ConfigOut"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "config"
                ],
                "summary": "Update Config",
                "operationId": "update_config_api_v1_config_put",
                "parameters": [
                    {
                        "name": "x-api-key",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-Api-Key"
                        }
                    },
                    {
                        "name": "authorization",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Authorization"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ConfigIn"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ConfigOut"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/domains": {
            "get": {
                "tags": [
                    "config"
                ],
                "summary": "List Domains",
                "operationId": "list_domains_api_v1_domains_get",
                "parameters": [
                    {
                        "name": "x-api-key",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-Api-Key"
                        }
                    },
                    {
                        "name": "authorization",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Authorization"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/DomainMappingOut"
                                    },
                                    "title": "Response List Domains Api V1 Domains Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "config"
                ],
                "summary": "Create Domain",
                "operationId": "create_domain_api_v1_domains_post",
                "parameters": [
                    {
                        "name": "x-api-key",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-Api-Key"
                        }
                    },
                    {
                        "name": "authorization",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Authorization"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DomainMappingIn"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DomainMappingOut"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ConfigIn": {
                "properties": {
                    "email_format": {
                        "type": "string",
                        "title": "Email Format"
                    },
                    "email_custom_pattern": {
                        "type": "string",
                        "title": "Email Custom Pattern",
                        "default": "{f}{last}"
                    },
                    "default_domain": {
                        "type": "string",
                        "title": "Default Domain"
                    },
                    "callback_enabled": {
                        "type": "boolean",
                        "title": "Callback Enabled",
                        "default": false
                    },
                    "callback_url": {
                        "type": "string",
                        "title": "Callback Url",
                        "default": ""
                    },
                    "callback_method": {
                        "type": "string",
                        "title": "Callback Method",
                        "default": "POST"
                    },
                    "callback_auth_header_name": {
                        "type": "string",
                        "title": "Callback Auth Header Name",
                        "default": ""
                    },
                    "callback_auth_header_value": {
                        "type": "string",
                        "title": "Callback Auth Header Value",
                        "default": ""
                    },
                    "callback_body_template": {
                        "type": "string",
                        "title": "Callback Body Template",
                        "default": "{\"employeeId\": \"{external_employee_id}\", \"email\": \"{email}\"}"
                    }
                },
                "type": "object",
                "required": [
                    "email_format",
                    "default_domain"
                ],
                "title": "ConfigIn"
            },
            "ConfigOut": {
                "properties": {
                    "email_format": {
                        "type": "string",
                        "title": "Email Format"
                    },
                    "email_custom_pattern": {
                        "type": "string",
                        "title": "Email Custom Pattern",
                        "default": "{f}{last}"
                    },
                    "default_domain": {
                        "type": "string",
                        "title": "Default Domain"
                    },
                    "callback_enabled": {
                        "type": "boolean",
                        "title": "Callback Enabled",
                        "default": false
                    },
                    "callback_url": {
                        "type": "string",
                        "title": "Callback Url",
                        "default": ""
                    },
                    "callback_method": {
                        "type": "string",
                        "title": "Callback Method",
                        "default": "POST"
                    },
                    "callback_auth_header_name": {
                        "type": "string",
                        "title": "Callback Auth Header Name",
                        "default": ""
                    },
                    "callback_auth_header_value": {
                        "type": "string",
                        "title": "Callback Auth Header Value",
                        "default": ""
                    },
                    "callback_body_template": {
                        "type": "string",
                        "title": "Callback Body Template",
                        "default": "{\"employeeId\": \"{external_employee_id}\", \"email\": \"{email}\"}"
                    }
                },
                "type": "object",
                "required": [
                    "email_format",
                    "default_domain"
                ],
                "title": "ConfigOut"
            },
            "DomainMappingIn": {
                "properties": {
                    "company": {
                        "type": "string",
                        "minLength": 1,
                        "title": "Company"
                    },
                    "domain": {
                        "type": "string",
                        "minLength": 1,
                        "title": "Domain"
                    }
                },
                "type": "object",
                "required": [
                    "company",
                    "domain"
                ],
                "title": "DomainMappingIn"
            },
            "DomainMappingOut": {
                "properties": {
                    "company": {
                        "type": "string",
                        "minLength": 1,
                        "title": "Company"
                    },
                    "domain": {
                        "type": "string",
                        "minLength": 1,
                        "title": "Domain"
                    },
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    }
                },
                "type": "object",
                "required": [
                    "company",
                    "domain",
                    "id"
                ],
                "title": "DomainMappingOut"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}